requirements.txt
fastapi==0.127.0
uvicorn==0.40.0
openai==2.14.0
pydantic==2.12.5
python-dotenv==1.2.1
# RAG dependencies - choose based on your implementation
chromadb==1.4.0
pytest==9.0.2
datasets==4.4.2
ragas==0.4.2
sentence-transformers==5.2.0
langchain-openai==1.1.6 # for eval
numpy==2.4.0
# faiss-cpu>=1.7.0